begin separating application testing from e2e
