+++ src/main/scala/org/constellation/ConstellationApp.scala
@@ -21,9 +21,9 @@ class ConstellationApp(
               schemaStr: String,
               channelName: String = s"channel_${channelIdToChannel.keys.size + 1}"
             )(implicit ec: ExecutionContext) = {
-    val response = clientApi.postNonBlocking[Some[ChannelOpenResponse]]("channel/open", ChannelOpenRequest(channelName, jsonSchema = Some(schemaStr)), timeout = 15 seconds)
-    response.map { resp: Option[ChannelOpenResponse] =>
-      logger.info(s"ChannelOpenResponse: ${resp.toString}")
+    val response = clientApi.postNonBlocking[Some[ChannelOpenResponse]]("channel/open", ChannelOpen(channelName, jsonSchema = Some(schemaStr)), timeout = 15 seconds)
+    response.map { resp =>
+//      logger.info(s"ChannelOpenResponse: ${resp.toString}")
       val channelMsg = resp.map { msg =>
         Channel(msg.genesisHash, channelName, msg)
       }
